refactor: top repos active
